anomaly-detection/services/feature-engineer/requirements.txt
faust-streaming==0.11.3
aiokafka==0.10.0
redis==5.0.1
kafka-python==2.0.2
numpy==1.26.0
prometheus-client==0.17.1
aiohttp==3.9.1
